Move 'src/domain/app' to 'src/app'
