+++ src/app/App.tsx
@@ -3,18 +3,18 @@ import React from 'react';
 import { BrowserRouter as Router, Route, Switch, Redirect } from 'react-router-dom';
 
 import apolloClient from './apolloClient';
-import CustomersPage from '../customers/CustomerPageContainer';
-import HarborsPage from '../harbors/HarborsPageContainer';
-import IndividualHarborPage from '../individualHarbor/IndividualHarborPageContainer';
-import IndividualCustomerPage from '../individualCustomer/IndividualCustomerPageContainer';
-import OfferPage from '../offer/OfferPageContainer';
-import LoginPage from '../auth/loginPage/LoginPage';
-import Page from '../page/Page';
-import PrivateRoute from '../auth/privateRoute/PrivateRoute';
-import ApplicationsPage from '../applications/ApplicationsPageContainer';
-import IndividualApplicationPage from '../individualApplication/IndividualApplicationPageContainer';
-import PricingPage from '../pricing/PricingPageContainer';
-import CallbackPage from '../auth/callbackPage/CallbackPage';
+import CustomersPage from '../domain/customers/CustomerPageContainer';
+import HarborsPage from '../domain/harbors/HarborsPageContainer';
+import IndividualHarborPage from '../domain/individualHarbor/IndividualHarborPageContainer';
+import IndividualCustomerPage from '../domain/individualCustomer/IndividualCustomerPageContainer';
+import OfferPage from '../domain/offer/OfferPageContainer';
+import LoginPage from '../domain/auth/loginPage/LoginPage';
+import Page from '../domain/page/Page';
+import PrivateRoute from '../domain/auth/privateRoute/PrivateRoute';
+import ApplicationsPage from '../domain/applications/ApplicationsPageContainer';
+import IndividualApplicationPage from '../domain/individualApplication/IndividualApplicationPageContainer';
+import PricingPage from '../domain/pricing/PricingPageContainer';
+import CallbackPage from '../domain/auth/callbackPage/CallbackPage';
 
 const App: React.FC = () => {
   return (
