+++ src/app/app.module.scss
